Embed the order profiles file in TestNewOrder.
The test reads a test file and uses the content of that file in the assertions. Create the file on the fly so that it's clear where the values for the assertions come from. It also prevents changes in the submitted order profiles test file from affecting TestNewOrder, which would be counter intuitive.

PiperOrigin-RevId: 310138143