Implement the tracking of `Order`s associated with `Encounter`s.
This change adds orders generated by `Order` and `Results` events to the relevant encounter. For each `Order` event, two cases are considered:
1. We are inside an ongoing `Encounter`. In this case, we simply need to update the encounter's status to "in-progress".
2. We are not inside an ongoing `Encounter` - either because no encounters exist, or because the most recent encounter has ended (status "finished" or "cancelled"). In this case we create an `Encounter` for just this `Order`.

In the `Results` step we need to consider one additional scenario. The `Order` associated with this result may or may not already exist in an `Encounter` that finished prior in the pathway. If one does exist, we don't need to do anything with the `Encounter`. If one does not exist, however, we need to follow the same logic as above.

Note that if an existing `Order` in a finished `Encounter` is being updated, the `ReportedDateTime` will be greater than the `Encounter`'s end time, which is fine since it's considered a correction anyway.

PiperOrigin-RevId: 316428219