Remove resource duplicates for `Practitioner` and `Location`.
Previously, it was possible to get duplicates of these resources since multiple other resources could refer to them. We maintain a map to track which we have already visited, and if so we return only the reference and do not add it again to the bundle. This is not necessary for other resources because they will never be referenced more than once at most.

PiperOrigin-RevId: 320971134
Change-Id: I607f513b535c6ab80cfb6457bdfa958fd9266f1a