Add datatypes DTM and SNM, added in HL7v2 > 2.5.1
Add a NewNUL method that is useful when we start having NUL types in tests.

DTM is essentially a TS. Test cases are the same, except for the additional precision that TS supports, and DTM does not.

SNM is supposed to be a phone number (digits and the "+" character) but it's assumed to be a string for now for simplicity.

PiperOrigin-RevId: 518212208
Change-Id: Ie2c73875c7a0b67cc6100178dd00d8ecddb1f275